Pass on doc strings, namespace usage and some Scrutinizer fixes

+++ iveeCore/Type.php
@@ -9,7 +9,6 @@
  * @author   Aineko Macx <ai@sknop.net>
  * @license  https://github.com/aineko-m/iveeCore/blob/master/LICENSE GNU Lesser General Public License
  * @link     https://github.com/aineko-m/iveeCore/blob/master/iveeCore/Type.php
- *
  */
 
 namespace iveeCore;
@@ -23,7 +22,6 @@ namespace iveeCore;
  * @author   Aineko Macx <ai@sknop.net>
  * @license  https://github.com/aineko-m/iveeCore/blob/master/LICENSE GNU Lesser General Public License
  * @link     https://github.com/aineko-m/iveeCore/blob/master/iveeCore/Type.php
- *
  */
 class Type extends SdeType
 {
@@ -59,7 +57,7 @@ class Type extends SdeType
     protected $portionSize;
 
     /**
-     * @var int $basePrice the base price of this Type.
+     * @var float $basePrice the base price of this Type.
      */
     protected $basePrice;
 
@@ -104,7 +102,7 @@ class Type extends SdeType
     }
 
     /**
-     * Loads all type names from DB to PHP
+     * Loads all type names from DB to PHP.
      *
      * @return void
      */
@@ -301,7 +299,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets all necessary data from SQL
+     * Gets all necessary data from SQL.
      *
      * @return array with attributes queried from DB
      * @throws \iveeCore\Exceptions\TypeIdNotFoundException when a typeID is not found
@@ -340,7 +338,7 @@ class Type extends SdeType
     }
 
     /**
-     * Sets attributes from SQL result row to object
+     * Sets attributes from SQL result row to object.
      *
      * @param array $row data from DB
      *
@@ -353,7 +351,7 @@ class Type extends SdeType
         $this->name        = $row['typeName'];
         $this->volume      = (float) $row['volume'];
         $this->portionSize = (int) $row['portionSize'];
-        $this->basePrice   = (int) $row['basePrice'];
+        $this->basePrice   = (float) $row['basePrice'];
         if (isset($row['marketGroupID']))
             $this->marketGroupID = (int) $row['marketGroupID'];
         if (isset($row['reprocessingSkillID']))
@@ -361,7 +359,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets the groupID of the Type
+     * Gets the groupID of the Type.
      *
      * @return int
      */
@@ -371,7 +369,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets the categoryID of the Type
+     * Gets the categoryID of the Type.
      *
      * @return int
      */
@@ -381,7 +379,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets the volume of the Type
+     * Gets the volume of the Type.
      *
      * @return float
      */
@@ -391,7 +389,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets the portion size of the Type (relevant for manufacturing and reprocessing)
+     * Gets the portion size of the Type (relevant for manufacturing and reprocessing).
      *
      * @return int
      */
@@ -404,7 +402,7 @@ class Type extends SdeType
      * Gets the base price.
      * Not to be confused with the base price calculated in industry activities for cost calculation.
      *
-     * @return int
+     * @return float
      */
     public function getBasePrice()
     {
@@ -412,7 +410,7 @@ class Type extends SdeType
     }
 
     /**
-     * Gets marketGroupID
+     * Gets marketGroupID.
      *
      * @return int marketGroupID
      */
@@ -422,7 +420,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns boolean on whether item can be sold/bought or not
+     * Returns boolean on whether item can be sold/bought or not.
      *
      * @return bool
      */
@@ -432,7 +430,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns the GlobalPriceData object for this Type
+     * Returns the GlobalPriceData object for this Type.
      *
      * @return \iveeCore\GlobalPriceData
      * @throws \iveeCore\Exceptions\NoPriceDataAvailableException if there is no price data available for this Type
@@ -444,7 +442,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns the RegionMarketData object for this Type
+     * Returns the RegionMarketData object for this Type.
      *
      * @param int $regionID of the region to get market data for. If none passed, default is looked up.
      *
@@ -459,7 +457,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns whether this Type can be reprocessed or not
+     * Returns whether this Type can be reprocessed or not.
      *
      * @return bool
      */
@@ -471,7 +469,7 @@ class Type extends SdeType
     /**
      * Returns the materials for Type. Since Crius this is only relevant for reprocessing.
      *
-     * @return array typeID => quantity
+     * @return int[] in the form typeID => quantity
      */
     public function getMaterials()
     {
@@ -482,7 +480,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns the specific reprocessing skill ID for this Type
+     * Returns the specific reprocessing skill ID for this Type.
      *
      * @return int
      */
@@ -492,7 +490,7 @@ class Type extends SdeType
     }
 
     /**
-     * Returns a MaterialMap object representing the reprocessing materials of the item
+     * Returns a MaterialMap object representing the reprocessing materials of the item.
      *
      * @param int $batchSize number of items being reprocessed, needs to be multiple of portionSize
      * @param float $equipmentYield the reprocessing yield of the station or array
@@ -545,7 +543,7 @@ class Type extends SdeType
     }
 
     /**
-     * Calculates the tax factor for reprocessing in stations (5% tax = factor of 0.95)
+     * Calculates the tax factor for reprocessing in stations (5% tax = factor of 0.95).
      *
      * @param float $standings with the corporation of the station you are reprocessing at
      *
